# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SecureGate API
  version: 2.0.0
  description: |
    REST API for **SecureGate** — a Smart Contractor & Visitor Access Management
    System for manufacturing facilities.

    ## Authentication

    There are three separate auth mechanisms, all cookie- or header-based. None
    of them use an `Authorization: Bearer <user token>` header, so Swagger UI's
    "Authorize" button only covers the cron endpoint.

    | Mechanism | Applies to | How it is obtained |
    |---|---|---|
    | `staff_token` cookie (JWT, HS256, 8h) | everything under `/api/staff/*` | `POST /api/staff/login` |
    | `track_session_{trackingToken}` cookie (1h) | `/api/visit`, `/api/track/document/reupload` | `POST /api/track/verify` |
    | `Authorization: Bearer $CRON_SECRET` | `/api/cron/document-expiry` | environment variable |

    Because the staff and tracking sessions are **httpOnly cookies**, Swagger UI
    cannot set them from JavaScript. To exercise those endpoints from Swagger UI:

    1. Call `POST /api/staff/login` (or `POST /api/track/verify`) from the same
       browser — the browser stores the cookie automatically.
    2. Serve this spec from the **same origin** as the API so the cookie is sent.

    Otherwise use `curl -c/-b` or Postman, which handle cookie jars directly.

    ## Roles

    `/api/staff/*` endpoints are gated per-role. Each operation documents the
    roles it accepts; a caller with a valid session but the wrong role also
...

    ## Conventions

...

...
servers:
...
paths:
...
  /api/track/verify:
    post:
      tags: [Public — Tracking]
      summary: Verify email ownership for a tracking token
      description: |
        Proves the caller owns the email behind a registration, then issues a
        `track_session_{token}` httpOnly cookie valid for **1 hour**. That cookie
        is what authorises `POST /api/visit` and
        `POST /api/track/document/reupload`.

        Note the asymmetry: `token` here is the **raw** `trackingToken`, whereas
        the value in `/track-status/{url}` URLs is encrypted.

        Navigating to the tracking search page clears every `track_session_*`
        cookie, so each registration must be re-verified per visit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, email]
              properties:
                token:
                  type: string
                  description: Raw (unencrypted) tracking token.
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verified. Sets the `track_session_{token}` cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
              example: track_session_abc123=budi%40majujaya.co.id; Path=/; HttpOnly; SameSite=Strict; Max-Age=3600
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Email does not match this registration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/staff/registrations/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Staff — Registrations]
      summary: Get a registration with its active documents
      description: |
        Roles: `HSE_ADMIN`, `HR_ADMIN`.

        Only `isActive: true` documents are included. Superseded rows remain in
        the table for history but are filtered out, otherwise a re-uploaded type
        would appear twice.
      security:
        - staffSessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration:
                    $ref: '#/components/schemas/RegistrationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Staff — Registrations]
      summary: Approve or reject a registration
      description: |
        Roles: `HSE_ADMIN`, `HR_ADMIN`.

        Only a `PENDING` registration can be reviewed — anything else returns
        **409**. There is deliberately **no un-reject path**: a rejected
        applicant either replaces documents via
        `POST /api/track/document/reupload` (which returns the record to
        `PENDING`) or registers again, which deletes the rejected row.

        `APPROVE` additionally requires every active document to be verified.

        Writes `REGISTRATION_APPROVED` / `REGISTRATION_REJECTED` to the audit
        log and emails the applicant.
      security:
        - staffSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [APPROVE, REJECT]
                rejectionReason:
                  type: string
                  description: Required when `action=REJECT`.
      responses:
        '200':
          description: Reviewed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration:
                    type: object
                    properties:
                      id:
                        type: string
                      status:
                        $ref: '#/components/schemas/RegistrationStatus'
                      rejectionReason:
                        type: string
                        nullable: true
                      fullName:
                        type: string
        '400':
          description: Invalid action, or `REJECT` without a reason.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already reviewed, or unverified documents block approval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Verify all documents before approving: SIO, SIA'

  # --------------------------------------------------------- STAFF DOCUMENTS -
...
components:
  securitySchemes:
    staffSessionCookie:
      type: apiKey
      in: cookie
      name: staff_token
      description: |
        HS256 JWT, 8-hour expiry, set by `POST /api/staff/login`. **httpOnly**,
        so Swagger UI cannot attach it programmatically — log in from the same
        browser and serve this spec from the same origin, or use curl/Postman.

    trackSessionCookie:
      type: apiKey
      in: cookie
      name: track_session_{trackingToken}
      description: |
        Per-registration session cookie, 1-hour expiry, set by
        `POST /api/track/verify`. The cookie **name** embeds the raw tracking
        token, so each registration carries its own cookie — this is why it
        cannot be expressed as a fixed name here. **httpOnly.**

    cronSecret:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer $CRON_SECRET`. Used only by `/api/cron/document-expiry`.'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: CUID.
  responses:
    BadRequest:
      description: Missing or invalid request parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: |
        No valid session, **or** a valid session whose role is not permitted —
        the implementation does not distinguish between the two.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Unauthorized
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ----- enums (mirrors prisma/schema.prisma) -----
    RegistrationType:
      type: string
      enum: [CONTRACTOR, VISITOR]
    RegistrationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    DocumentType:
      type: string
      enum: [KTP, BPJS, SIO, SIA, FACE_PHOTO]
...
    RegistrationListItem:
      type: object
      properties:
        id:
          type: string
        trackingToken:
          type: string
          description: Raw token (this endpoint is staff-only).
        type:
          $ref: '#/components/schemas/RegistrationType'
        fullName:
          type: string
        company:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RegistrationStatus'
        createdAt:
          type: string
          format: date-time
    RegistrationDetail:
      allOf:
        - $ref: '#/components/schemas/RegistrationListItem'
        - type: object
          properties:
            rejectionReason:
              type: string
              nullable: true
            photoPath:
              type: string
              nullable: true
            documents:
              type: array
              description: Active documents only.
              items:
                $ref: '#/components/schemas/Document'
    Document:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/DocumentType'
        filePath:
          type: string
          description: Blob pathname. Stream it via `/api/staff/documents/{id}/view`.
        expiryDate:
          type: string
          format: date-time
          nullable: true
        isVerified:
          type: boolean
...